Split consensus between Follower and Leader
